% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steprfAVIPredictors.R
\name{steprfAVIPredictors}
\alias{steprfAVIPredictors}
\title{Extract names of the selected predictive variables by steprf}
\usage{
steprfAVIPredictors(steprf1, trainx)
}
\arguments{
\item{steprf1}{a list of output of 'steprf' function.}

\item{trainx}{a dataframe or matrix contains columns of predictor variables.}
}
\value{
A list with the following components:
1) variables.most.accurate: a list of predictive variables contained in the
 most accurate RF model,  2) PABV: a list of predictive variables with positive
  contributions to the predictive accuracy of RF models, that is,
   predictive accuracy boosting variable (PABV), 3) PARV: a list of
   predictive variables with negative contributions to the predictive
   accuracy of RF models, that is, predictive accuracy reducing variable,
   and 4) max.predictive.accuracy: the predictive accuracy of the most
   accurate RF model.
}
\description{
This function is to extract names of the selected predictive variables by steprfAVI.
}
\examples{
\donttest{
library(spm)
data(petrel)
set.seed(1234)
steprf1 <- steprfAVI(trainx = petrel[, c(1,2, 6:9)], trainy = petrel[, 5],
 nsim = 10, min.n.var = 2)

steprfAVIPredictors(steprf1, trainx = petrel[, c(1,2, 6:9)])
}

}
\references{
Li, J. (2022). Spatial Predictive Modeling with R. Boca Raton,
Chapman and Hall/CRC.

Li, J. (2019). "A critical review of spatial predictive modeling
process in environmental sciences with reproducible examples in R." Applied
Sciences 9: 2048.

Li, J., Siwabessy, J., Huang, Z., and Nichol, S. (2019). "Developing an
optimal spatial predictive model for seabed sand content using machine
learning, geostatistics and their hybrid methods." Geosciences 9 (4):180.
}
\author{
Jin Li
}
